refactor: switch to auth module 'github.com/Tibz-Dankan/reserve-now-microservices/auth'
